+++ www/SpeechRecognitionEvent.js
@@ -1,8 +1,9 @@
-var SpeechRecognitionEvent = function() {
-    this.resultIndex;
-    this.results;
-    this.interpretation;
-    this.emma;
+var SpeechRecognitionEvent = function () {
+    this.type = "result";
+    this.resultIndex = 0;
+    this.results = new SpeechRecognitionResultList();
+    this.interpretation = null;
+    this.emma = null;
 };
 
 module.exports = SpeechRecognitionEvent;