+++ www/SpeechRecognitionResult.js
@@ -1,12 +1,16 @@
 // A complete one-shot simple response
-var SpeechRecognitionResult = function() {
-    this._result = [];
-    this.length = 0;
-    this.final = false;
+// An extension of the Array object to also operate similarly to Chrome's implementation.
+var SpeechRecognitionResult = function () {
+    Array.call(this);
+
+    this.isFinal = false;
 };
 
-SpeechRecognitionResult.prototype.item = function(item) {
-    return this._result[item];
+SpeechRecognitionResult.prototype = new Array;
+SpeechRecognitionResult.prototype.constructor = SpeechRecognitionResult;
+
+SpeechRecognitionResult.prototype.item = function (item) {
+    return this[item];
 };
 
 module.exports = SpeechRecognitionResult;